+++ rs/web-transport-noq/src/server.rs
@@ -49,7 +49,7 @@ impl ServerBuilder {
     /// Enable the specified congestion controller.
     pub fn with_congestion_control(mut self, algorithm: CongestionControl) -> Self {
         self.congestion_controller = match algorithm {
-            CongestionControl::LowLatency => Some(Arc::new(noq::congestion::BbrConfig::default())),
+            CongestionControl::LowLatency => Some(Arc::new(noq::congestion::Bbr3Config::default())),
             // TODO BBR is also higher throughput in theory.
             CongestionControl::Throughput => {
                 Some(Arc::new(noq::congestion::CubicConfig::default()))
@@ -175,6 +175,11 @@ impl Request {
     pub async fn reject(self, status: http::StatusCode) -> Result<(), ServerError> {
         self.connect.reject(status).await?;
         Ok(())
+    }
+
+    /// Returns the underlying QUIC connection.
+    pub fn conn(&self) -> &noq::Connection {
+        &self.conn
     }
 
     /// Returns the CONNECT request that was sent by the client.