+++ rs/web-transport-noq/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [0.1.1](https://github.com/moq-dev/web-transport/compare/web-transport-noq-v0.1.0...web-transport-noq-v0.1.1) - 2026-05-24
+
+### Other
+
+- release ([#228](https://github.com/moq-dev/web-transport/pull/228))
+
+## [0.1.0](https://github.com/moq-dev/web-transport/compare/web-transport-noq-v0.0.4...web-transport-noq-v0.1.0) - 2026-05-21
+
+### Other
+
+- update iroh and noq to 1.0-rc.0 ([#236](https://github.com/moq-dev/web-transport/pull/236))
+- update to iroh 0.98 and noq 0.18 ([#230](https://github.com/moq-dev/web-transport/pull/230))
+
+## [0.0.4](https://github.com/moq-dev/web-transport/compare/web-transport-noq-v0.0.3...web-transport-noq-v0.0.4) - 2026-04-07
+
+### Other
+
+- Expose conn() by reference and fix Python bindings ([#227](https://github.com/moq-dev/web-transport/pull/227))
+
 ## [0.0.2](https://github.com/moq-dev/web-transport/compare/web-transport-noq-v0.0.1...web-transport-noq-v0.0.2) - 2026-03-11
 
 ### Other